Prepping for VS2017RC compat; updating based on VSTS guidance
