+++ VSTSExt/VSTSExt.csproj
@@ -1,12 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
-    <MinimumVisualStudioVersion>14.0</MinimumVisualStudioVersion>
+    <MinimumVisualStudioVersion>15.0</MinimumVisualStudioVersion>
     <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
     <FileUpgradeFlags>
     </FileUpgradeFlags>
     <OldToolsVersion>14.0</OldToolsVersion>
-    <UpgradeBackupLocation />
+    <UpgradeBackupLocation>
+    </UpgradeBackupLocation>
     <PublishUrl>publish\</PublishUrl>
     <Install>true</Install>
     <InstallFrom>Disk</InstallFrom>
@@ -76,6 +77,7 @@
     <None Include="AssemblyInfo.cs" />
     <None Include="package.json" />
     <None Include="settings.tfx.json" />
+    <None Include="tsconfig.json" />
     <None Include="VSTSExtension.csproj">
       <SubType>Designer</SubType>
     </None>
@@ -103,21 +105,16 @@
     <Content Include=".gitignore">
       <IncludeInVSIX>true</IncludeInVSIX>
     </Content>
-    <Content Include="css\app.css" />
+    <Content Include="static\css\app.css" />
     <Content Include="gruntfile.js" />
-    <Content Include="img\logo.PNG" />
-    <Content Include="index.html" />
-    <Content Include="scripts\VSS.SDK.js" />
+    <Content Include="static\images\logo.PNG" />
+    <Content Include="static\index.html" />
+    <Content Include="src\VSS.SDK.js" />
     <Content Include="test\TestSpec.js" />
     <Content Include="VSOE.ico" />
   </ItemGroup>
   <ItemGroup>
-    <TypeScriptCompile Include="scripts\app.ts" />
-    <TypeScriptCompile Include="typings\jquery\jquery.d.ts" />
-    <TypeScriptCompile Include="typings\knockout\knockout.d.ts" />
-    <TypeScriptCompile Include="typings\q\Q.d.ts" />
-    <TypeScriptCompile Include="typings\requirejs\require.d.ts" />
-    <TypeScriptCompile Include="typings\tsd.d.ts" />
+    <TypeScriptCompile Include="src\my-module\app.ts" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
