+++ VSTSExt/package.json
@@ -1,13 +1,24 @@
 {
+  "scripts": {
+    "build": "tsc -p .",
+    "postbuild": "npm run package",
+    "package": "tfx extension create",
+    "clean": "rimraf ./dist && rimraf ./*.vsix"
+  },
   "devDependencies": {
     "grunt": "^1.0.0",
     "grunt-contrib-copy": "*",
     "grunt-contrib-jasmine": "*",
     "grunt-exec": "*",
     "jasmine": "^2.5.1",
-    "tfx-cli": "^0.3.30",
-    "typings": "^1.3.3",
-    "vss-web-extension-sdk": "^1.98.0"
+    "rimraf": "^2.5.4",
+    "tfx-cli": "^0.3.45",
+    "typescript": "^2.1.4"
+  },
+  "dependencies": {
+    "@types/jquery": "^2.0.34",
+    "@types/q": "0.0.32",
+    "vss-web-extension-sdk": "^2.109.1"
   },
   "name": "",
   "private": true,
