--- VSTSExt/scripts/app.ts
@@ -1,28 +1,0 @@
-﻿/// <reference path='../node_modules/vss-web-extension-sdk/typings/VSS.d.ts' />
-
-class Greeter {
-    element: HTMLElement;
-    span: HTMLElement;
-    timerToken: number;
-
-    constructor(element: HTMLElement) {
-        this.element = element;
-        this.element.innerHTML += "The time is: ";
-        this.span = document.createElement('span');
-        this.element.appendChild(this.span);
-        this.span.innerText = new Date().toUTCString();
-    }
-
-    start() {
-        this.timerToken = setInterval(() => this.span.innerHTML = new Date().toUTCString(), 500);
-    }
-
-    stop() {
-        clearTimeout(this.timerToken);
-    }
-
-}
-
-var el = document.getElementById('content');
-var greeter = new Greeter(el);
-greeter.start();
