+++ VSTSExt/src/app.js
@@ -1,0 +1,19 @@
+/// <reference types="vss-web-extension-sdk" />
+class Greeter {
+    constructor(element) {
+        this.element = element;
+        this.element.innerHTML += "The time is: ";
+        this.span = document.createElement('span');
+        this.element.appendChild(this.span);
+        this.span.innerText = new Date().toUTCString();
+    }
+    start() {
+        this.timerToken = setInterval(() => this.span.innerHTML = new Date().toUTCString(), 500);
+    }
+    stop() {
+        clearTimeout(this.timerToken);
+    }
+}
+var el = document.getElementById('content');
+var greeter = new Greeter(el);
+greeter.start();
