--- VSTSExt/typings/tsd.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="../node_modules/vss-web-extension-sdk/typings/tfs.d.ts" />
-/// <reference path="../node_modules/vss-web-extension-sdk/typings/vss.d.ts" />
