+++ VstsExtProjTemplates.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.26020.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSTSExtensionPackage", "VSTSExtensionPackage\VSTSExtensionPackage.csproj", "{EC7CC6D8-04F1-4A5E-872D-72DF5967B7ED}"
 EndProject
